[skip ci] changes based on merge conflict updates
