+++ src/fidesapi/routes/util.py
@@ -1,3 +1,6 @@
+import zlib
+from base64 import urlsafe_b64decode as b64d
+from base64 import urlsafe_b64encode as b64e
 from pathlib import Path
 
 from fastapi import APIRouter
@@ -17,3 +20,13 @@ def get_resource_type(router: APIRouter) -> str:
         The router's resource type
     """
     return router.prefix.replace(f"{API_PREFIX}/", "", 1)
+
+
+def obscure(data: bytes) -> bytes:
+    "obscures data as a minor security measure"
+    return b64e(zlib.compress(data, 9))
+
+
+def unobscure(obscured: bytes) -> bytes:
+    "unobscures data as a minor security measure"
+    return zlib.decompress(b64d(obscured))
