+++ src/fidesapi/routes/visualize.py
@@ -11,7 +11,6 @@ from fideslang import model_map
 from fidesapi.routes.crud import list_resource
 from fidesapi.routes.util import API_PREFIX, get_resource_type
 from fidesapi.sql_models import sql_model_map
-from fidesapi.utils.helpers import get_resource_type
 from fidesctl.core import visualize
 
 # pylint: disable=redefined-outer-name,cell-var-from-loop
